refactor(stories): split story detail into section components
