+++ client/src/containers/stories/detail/index.tsx
@@ -4,7 +4,7 @@ import { useEffect } from "react";
 
 import Image from "next/image";
 import { useLocale, useTranslations } from "next-intl";
-import { ArrowLeft, CirclePlay, ExternalLink, FileText } from "lucide-react";
+import { ArrowLeft } from "lucide-react";
 import { useMap } from "react-map-gl/mapbox";
 import { useAtomValue } from "jotai";
 
@@ -13,64 +13,12 @@ import { useGetStories } from "@/types/generated/story";
 import { DEFAULT_LOCALE } from "@/i18n/routing";
 import RichText from "@/components/ui/rich-text";
 import { sidebarOpenAtom, useSyncSearchParams } from "@/store/map";
-import type { DefaultFurtherInfoComponent } from "@/types/generated/strapi.schemas";
 import { CMS_MEDIA_BASE } from "@/lib/cms";
-
-type StoryDatasetAttributes = {
-  slug?: string;
-  title?: string;
-  layers?: { layer?: { data?: { attributes?: { slug?: string } } } }[];
-};
+import FurtherInfo from "./further-info";
+import RelatedDatasets, { type StoryDatasetAttributes } from "./related-datasets";
 
 type StoryDetailProps = {
   slug: string;
-};
-
-const FurtherInfoItem = ({
-  item,
-  locale,
-}: {
-  item: DefaultFurtherInfoComponent;
-  locale: string;
-}) => {
-  const content =
-    locale !== DEFAULT_LOCALE
-      ? (item[`content_${locale}` as keyof DefaultFurtherInfoComponent] as string | undefined) ??
-        item.content
-      : item.content;
-
-  const safeUrl = /^(https?:|mailto:)/i.test((item.url ?? "").trim()) ? item.url : undefined;
-
-  const icon =
-    item.type === "video" ? (
-      <CirclePlay className="h-4 w-4 shrink-0" />
-    ) : item.type === "paper" ? (
-      <FileText className="h-4 w-4 shrink-0" />
-    ) : (
-      <ExternalLink className="h-4 w-4 shrink-0" />
-    );
-
-  const inner = (
-    <>
-      <span className="mt-0.5">{icon}</span>
-      <span>{content ?? item.url}</span>
-    </>
-  );
-
-  if (safeUrl) {
-    return (
-      <a
-        href={safeUrl}
-        target="_blank"
-        rel="noopener noreferrer"
-        className="flex items-start gap-2 text-sm underline underline-offset-2 hover:text-green-light"
-      >
-        {inner}
-      </a>
-    );
-  }
-
-  return <span className="flex items-start gap-2 text-sm">{inner}</span>;
 };
 
 const StoryDetail = ({ slug }: StoryDetailProps) => {
@@ -168,33 +116,9 @@ const StoryDetail = ({ slug }: StoryDetailProps) => {
           </div>
         )}
 
-        {(attributes?.further_information?.length ?? 0) > 0 && (
-          <div className="space-y-2">
-            <h2 className="text-sm font-semibold uppercase tracking-wide">
-              {t("Further information")}
-            </h2>
-            <div className="space-y-2">
-              {attributes?.further_information?.map((item, i) => (
-                <FurtherInfoItem key={item.id ?? i} item={item} locale={locale} />
-              ))}
-            </div>
-          </div>
-        )}
+        <FurtherInfo items={attributes?.further_information ?? []} locale={locale} />
 
-        {storyDatasets.length > 0 && (
-          <div className="space-y-2">
-            <h2 className="text-sm font-semibold uppercase tracking-wide">
-              {t("Related datasets")}
-            </h2>
-            <ul className="space-y-1">
-              {storyDatasets.map((d) => (
-                <li key={d.id} className="text-sm">
-                  {d.attributes?.title ?? d.attributes?.slug}
-                </li>
-              ))}
-            </ul>
-          </div>
-        )}
+        <RelatedDatasets datasets={storyDatasets} />
       </div>
     </div>
   );
